+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_commands.py
@@ -5744,3 +5744,42 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         # delete
         self.cmd(
             'aks delete -g {resource_group} -n {name} --yes --no-wait', checks=[self.is_empty()])
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
+    def test_aks_create_and_update_with_http_proxy_config(self, resource_group, resource_group_location):
+        aks_name = self.create_random_name('cliakstest', 16)
+        self.kwargs.update({
+            'resource_group': resource_group,
+            'name': aks_name,
+            'kube_proxy_path': _get_test_data_file('kubeproxyconfig.json'),
+            'ssh_key_value': self.generate_ssh_keys()
+        })
+
+        create_cmd = 'aks create --resource-group={resource_group} --name={name} --ssh-key-value={ssh_key_value}' \
+                     '--kube-proxy-config={kube_proxy_path} --yes -o json'
+        self.cmd(create_cmd, checks=[
+            self.check('provisioningState', 'Succeeded'),
+        ])
+
+        self.cmd(create_cmd, checks=[
+            self.check('networkProfile.kubeProxyConfig.enabled','true'),
+            self.check('networkProfile.kubeProxyConfig.mode','IPTABLES'),
+        ])
+
+        self.kwargs.update({
+            'resource_group': resource_group,
+            'name': aks_name,
+            'kube_proxy_path': _get_test_data_file('kubeproxyconfig_update.json'),
+        })
+
+        update_cmd = 'aks update --resource-group={resource_group} --name={name} --kube-proxy-config={kube_proxy_path}'
+
+        self.cmd(update_cmd, checks=[
+            self.check('networkProfile.kubeProxyConfig.enabled','true'),
+            self.check('networkProfile.kubeProxyConfig.mode','IPVS'),
+            self.check('networkProfile.kubeProxyConfig.ipvsConfig.Scheduler', 'LeastConnection'),
+            self.check('networkProfile.kubeProxyConfig.ipvsConfig.TCPTimeoutSeconds', '900'),
+            self.check('networkProfile.kubeProxyConfig.ipvsConfig.TCPFINTimeoutSeconds', '120'),
+            self.check('networkProfile.kubeProxyConfig.ipvsConfig.UDPTimeoutSeconds', '30s'),
+        ])